cast added in SPEA2 archive
git-svn-id: svn://scm.gforge.inria.fr/svnroot/paradiseo@1859 331e1502-861f-0410-8da2-ba01fb791d7f

+++ trunk/paradiseo-moeo/src/archive/moeoSPEA2Archive.h
@@ -162,7 +162,7 @@ public:
         //If the archive is empty, put in the best elements of the pop
         if (borne < maxSize)
         {
-            foo= std::min(_pop.size(), maxSize-borne);
+            foo= std::min((unsigned int)_pop.size(), maxSize-borne);
             for (i=0; i< foo ; i++)
             {
                 push_back(_pop[copy_pop[i].index]);
